+++ index.js
@@ -1,0 +1,5 @@
+const PrivacyCompliance = require('./src/privacy_compliance');
+
+module.exports = {
+  PrivacyCompliance,
+};