+++ CHANGELOG.md
@@ -1,4 +1,9 @@
-## 1.1 (June 29th, 2020)
+## 1.2.0 (June 30th, 2020)
 
 - Adds support for Generators
 - First generators support building the US Privacy String
+
+## 1.1.0 (June 29th, 2020)
+
+- Adds support for babel transpiled source
+- Adds logging for framework